Backups/zzz_arrums_PGM_EM_DENO/_ZZZ__ARRUMMSSS/_Backups_OK_implementacoes/backups_tentativas_ecommerce/backup_ecommerce_OK_talves_use_1/src/modules_entitys/FindProductByCep/FindProductByCep.ts
import { ProductRepositoryProtocol } from "@src/entitys/product/repository/contract/ProductRepository.contract"
import { FindProductByCepProtocol } from "@src/modules_entitys/FindProductByCep/FindProductByCepProtocol.contract"

export class FindProductByCep implements FindProductByCepProtocol {
  constructor(private readonly productRepository: ProductRepositoryProtocol) {}

  async execute(cep: string) {
    const arrayItems = await this.productRepository.getItems()
    const productFound = arrayItems.find((p) => {
      const ruleCep = p.productControl.cepInicial <= cep && p.productControl.cepFinal >= cep
      return ruleCep
    })
    return productFound
  }
}
